[feat] home Bottom nav 구현, cinemaHomeActivity Fragment로 변경
