+++ app/src/main/AndroidManifest.xml
@@ -18,25 +18,16 @@
             android:name=".presentation.selection.CinemaSelectionActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.home.CinemaHomeActivity"
-            android:exported="true"
-            android:screenOrientation="portrait">
+            android:name=".presentation.detail.MovieDetailActivity"
+            android:exported="false" />
+        <activity
+            android:name=".presentation.main.MainActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".presentation.detail.MovieDetailActivity"
-            android:exported="false" />
-        <activity
-            android:name=".presentation.main.MainActivity"
-            android:exported="false"/>
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
-<!--            </intent-filter>-->
     </application>
 
 </manifest>
