+++ app/src/main/java/com/sopt/lottecinemaaos/presentation/main/MainActivity.kt
@@ -3,10 +3,20 @@ package com.sopt.lottecinemaaos.presentation.main
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.sopt.lottecinemaaos.R
+import com.sopt.lottecinemaaos.databinding.ActivityMainBinding
+import com.sopt.lottecinemaaos.util.base.BindingActivity
 
-class MainActivity : AppCompatActivity() {
+class MainActivity : BindingActivity<ActivityMainBinding>(R.layout.activity_main) {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        clickFloatingBtn()
+    }
+
+    private fun clickFloatingBtn() {
+        with(binding) {
+            btnReserve.setOnClickListener {
+                bottomNavMain.selectedItemId = R.id.menu_reserve
+            }
+        }
     }
 }
