+++ app/src/main/res/drawable/ic_event.xml
@@ -1,0 +1,41 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0,0h24v24h-24z"/>
+    <path
+        android:strokeWidth="1"
+        android:pathData="M3.44,6.38H19.97V16.7C19.97,18.46 18.54,19.88 16.79,19.88H3.44V6.38Z"
+        android:fillColor="#00000000"
+        android:strokeColor="#000000"/>
+    <path
+        android:pathData="M11.71,14.58C12.157,14.58 12.52,14.217 12.52,13.77C12.52,13.323 12.157,12.96 11.71,12.96C11.263,12.96 10.9,13.323 10.9,13.77C10.9,14.217 11.263,14.58 11.71,14.58Z"
+        android:strokeLineJoin="round"
+        android:strokeWidth="0.75"
+        android:fillColor="#00000000"
+        android:strokeColor="#000000"
+        android:strokeLineCap="round"/>
+    <path
+        android:strokeWidth="1"
+        android:pathData="M11.71,6.38C11.71,6.38 13.2,3.96 15.59,4.02C17.98,4.08 17.07,6.33 15.42,6.38"
+        android:strokeLineJoin="round"
+        android:fillColor="#00000000"
+        android:strokeColor="#000000"
+        android:strokeLineCap="round"/>
+    <path
+        android:strokeWidth="1"
+        android:pathData="M11.71,6.38C11.71,6.38 10.22,3.96 7.84,4.02C5.46,4.08 6.36,6.33 8.01,6.38"
+        android:strokeLineJoin="round"
+        android:fillColor="#00000000"
+        android:strokeColor="#000000"
+        android:strokeLineCap="round"/>
+    <path
+        android:strokeWidth="1"
+        android:pathData="M11.71,6.38V11.8"
+        android:fillColor="#00000000"
+        android:strokeColor="#000000"/>
+  </group>
+</vector>
