+++ app/src/main/res/drawable/ic_mypage.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0,0h24v24h-24z"/>
+    <path
+        android:pathData="M11.73,8.55C12.404,8.55 12.95,8.004 12.95,7.33C12.95,6.656 12.404,6.11 11.73,6.11C11.056,6.11 10.51,6.656 10.51,7.33C10.51,8.004 11.056,8.55 11.73,8.55Z"
+        android:strokeLineJoin="round"
+        android:strokeWidth="1.12"
+        android:fillColor="#00000000"
+        android:strokeColor="#000000"
+        android:strokeLineCap="round"/>
+    <path
+        android:pathData="M3.44,19.66C3.44,19.66 3.81,10.41 11.72,10.41C19.63,10.41 19.83,19.66 19.83,19.66"
+        android:strokeLineJoin="round"
+        android:strokeWidth="1.12"
+        android:fillColor="#00000000"
+        android:strokeColor="#000000"
+        android:strokeLineCap="round"/>
+  </group>
+</vector>
