+++ app/src/main/res/drawable/ic_ticket_black.xml
@@ -1,0 +1,59 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="42dp"
+    android:height="42dp"
+    android:viewportWidth="42"
+    android:viewportHeight="42">
+  <path
+      android:pathData="M10,13.1H32"
+      android:strokeLineJoin="round"
+      android:strokeWidth="0.96"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M13.46,13.1V30.8H18.94C18.94,30.8 19.27,29.28 21.09,29.28C22.91,29.28 23.24,30.8 23.24,30.8H28.54V13.1H13.46Z"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M18.63,21.71V26.71"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M18.88,24.21H23.88"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M17.41,15.11H23.88"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M19.66,20.86H23.88"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M19.77,13.05L22.97,15.16"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M21.77,14.39L23.21,13"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"/>
+  <path
+      android:strokeWidth="1"
+      android:pathData="M20.19,16.83C20.19,16.83 19.66,19.23 21.77,19.19C23.77,19.15 23.1,16.83 23.1,16.83"
+      android:strokeLineJoin="round"
+      android:fillColor="#00000000"
+      android:strokeColor="#000000"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M18.25,21.45C18.598,21.45 18.88,21.168 18.88,20.82C18.88,20.472 18.598,20.19 18.25,20.19C17.902,20.19 17.62,20.472 17.62,20.82C17.62,21.168 17.902,21.45 18.25,21.45Z"
+      android:fillColor="#000000"/>
+</vector>
