+++ app/src/main/res/drawable/ic_ticket_red.xml
@@ -1,0 +1,84 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:aapt="http://schemas.android.com/aapt"
+    android:width="42dp"
+    android:height="42dp"
+    android:viewportWidth="42"
+    android:viewportHeight="42">
+  <group>
+    <clip-path
+        android:pathData="M0,0h42v42h-42z"/>
+    <path
+        android:strokeWidth="1"
+        android:pathData="M21,21m-20.5,0a20.5,20.5 0,1 1,41 0a20.5,20.5 0,1 1,-41 0"
+        android:strokeColor="#ffffff">
+      <aapt:attr name="android:fillColor">
+        <gradient 
+            android:startX="-0"
+            android:startY="3.405"
+            android:endX="45.992"
+            android:endY="9.776"
+            android:type="linear">
+          <item android:offset="0" android:color="#FFDB291D"/>
+          <item android:offset="1" android:color="#FFEC6680"/>
+        </gradient>
+      </aapt:attr>
+    </path>
+    <group>
+      <clip-path
+          android:pathData="M9,9h24v24h-24z"/>
+      <path
+          android:pathData="M10,12.2H32"
+          android:strokeLineJoin="round"
+          android:strokeWidth="0.96"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"
+          android:strokeLineCap="round"/>
+      <path
+          android:strokeWidth="1"
+          android:pathData="M13.46,12.2V29.9H18.94C18.94,29.9 19.27,28.38 21.09,28.38C22.91,28.38 23.24,29.9 23.24,29.9H28.54V12.2H13.46Z"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"
+          android:strokeLineCap="round"/>
+      <path
+          android:strokeWidth="1"
+          android:pathData="M18.63,20.81V25.81"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"/>
+      <path
+          android:strokeWidth="1"
+          android:pathData="M18.88,23.31H23.88"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"/>
+      <path
+          android:strokeWidth="1"
+          android:pathData="M17.41,14.21H23.88"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"/>
+      <path
+          android:strokeWidth="1"
+          android:pathData="M19.66,19.96H23.88"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"/>
+      <path
+          android:strokeWidth="1"
+          android:pathData="M19.77,12.15L22.97,14.26"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"/>
+      <path
+          android:strokeWidth="1"
+          android:pathData="M21.77,13.49L23.21,12.1"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"/>
+      <path
+          android:strokeWidth="1"
+          android:pathData="M20.19,15.93C20.19,15.93 19.66,18.33 21.77,18.29C23.77,18.25 23.1,15.93 23.1,15.93"
+          android:strokeLineJoin="round"
+          android:fillColor="#00000000"
+          android:strokeColor="#ffffff"
+          android:strokeLineCap="round"/>
+      <path
+          android:pathData="M18.25,20.55C18.598,20.55 18.88,20.268 18.88,19.92C18.88,19.572 18.598,19.29 18.25,19.29C17.902,19.29 17.62,19.572 17.62,19.92C17.62,20.268 17.902,20.55 18.25,20.55Z"
+          android:fillColor="#ffffff"/>
+    </group>
+  </group>
+</vector>
