+++ app/src/main/res/drawable/shape_red_fill_16_rect.xml
@@ -7,9 +7,7 @@
         android:color="@color/red_DB291D"
         />
     <padding
-        android:left="8dp"
         android:top="4dp"
-        android:right="8dp"
         android:bottom="4dp"
         />
 </shape>
