+++ app/src/main/res/values/themes.xml
@@ -2,7 +2,7 @@
     <!-- Base application theme. -->
     <style name="Theme.LotteCinemaAOS" parent="Theme.AppCompat.DayNight.NoActionBar">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/purple_500</item>
+        <item name="colorPrimary">@color/black</item>
         <item name="colorPrimaryVariant">@color/purple_700</item>
         <item name="colorOnPrimary">@color/white</item>
         <!-- Secondary brand color. -->
